Working on UI testing program
I'm following the article that I found on UI automation testing and i've
gotten to the point of accessing my program, but now I need to do some
research and figure out what the next thing I need to find is. I think
that I need to find the page that everything is in but i'm not sure. No
one else knows yet either so we'll see who finds the answer first over
the weekend.